Added authorization for users

+++ backend/routes/posts.js
@@ -42,7 +42,8 @@ router.post(
         const post = new Post({
             title: req.body.title,
             content: req.body.content,
-            imagePath: url + "/images/" + req.file.filename
+            imagePath: url + "/images/" + req.file.filename,
+            creator: req.userData.userId
         });
         post.save().then(createdPost => {
             res.status(201).json({
@@ -72,15 +73,23 @@ router.put(
             _id: req.body.id,
             title: req.body.title,
             content: req.body.content,
-            imagePath: imagePath
+            imagePath: imagePath,
+            creator: req.userData.userId
         });
         console.log(post);
         Post.updateOne({
-            _id: req.params.id
+            _id: req.params.id,
+            creator: req.userData.userId
         }, post).then(result => {
-            res.status(200).json({
-                message: "Update successful!"
-            });
+            if (result.nModified > 0) {
+                res.status(200).json({
+                    message: "Update successful!"
+                });
+            } else {
+                res.status(401).json({
+                    message: "Not authorized!"
+                });
+            }
         });
     }
 );
@@ -120,17 +129,19 @@ router.get("/:id", (req, res, next) => {
     });
 });
 
-router.delete("/:id",
-    checkAuth,
-    (req, res, next) => {
-        Post.deleteOne({
-            _id: req.params.id
-        }).then(result => {
-            console.log(result);
+router.delete("/:id", checkAuth, (req, res, next) => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+        if (result.n > 0) {
             res.status(200).json({
-                message: "Post deleted!"
+                message: "Post deleted successful!"
             });
-        });
+        } else {
+            res.status(401).json({
+                message: "Not authorized!"
+            });
+        }
+
     });
+});
 
 module.exports = router;
